Backend/requirements.txt
fastapi
uvicorn
openai
pytesseract
Pillow
transformers 
google-generativeai 